+++ IkaLog.py
@@ -22,6 +22,7 @@ import signal
 
 from ikalog.engine import *
 from IkaConfig import *
+from ikalog.utils import *
 
 def signal_handler(num, frame):
     IkaUtils.dprint('IkaLog: got signal %d' % num)
