+++ ikalog/outputs/__init__.py
@@ -1,1 +1,15 @@
-__author__ = 'mzsm'
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+from .console import IkaOutput_Console
+from .csv import IkaOutput_CSV
+from .csv_alivesquids import IkaOutput_CSV_AliveSquids
+from .debug import IkaOutput_DebugLog
+from .fluentd import IkaOutput_Fluentd
+from .hue import IkaOutput_Hue
+from .json import IkaOutput_JSON
+from .preview import IkaOutput_Screen
+from .screenshot import IkaOutput_Screenshot
+from .slack import IkaOutput_Slack
+from .twitter import IkaOutput_Twitter
+from .videorecorder import IkaOutput_OBS
+from .weapon_training import IkaOutput_WeaponTraining
