+++ ikalog/outputs/json.py
@@ -21,9 +21,9 @@
 from datetime import datetime
 import time
 import json
-import os
 
-from ikalog.IkaUtils import *
+from ikalog.utils import *
+
 
 # Needed in GUI mode
 try:
