+++ ikalog/outputs/slack.py
@@ -18,7 +18,7 @@
 #  limitations under the License.
 #
 
-from ikalog.IkaUtils import *
+from ikalog.utils import *
 
 # Needed in GUI mode
 try:
