+++ ikalog/scenes/game_finish.py
@@ -17,7 +17,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #
-from ..IkaUtils import *
+from ikalog.utils import *
 
 
 class IkaScene_GameFinish:
