+++ ikalog/scenes/lobby.py
@@ -17,7 +17,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #
-from ikalog.IkaUtils import *
+from ikalog.utils import *
 
 
 class IkaScene_Lobby:
